+++ Sources/Classes/core/View/MiniAppViewHandler.swift
@@ -874,7 +874,7 @@ extension MiniAppViewHandler: WKUIDelegate {
 
 // MARK: - Universal Bridge
 extension MiniAppViewHandler {
-    internal func sendJsonStringToMiniApp(string jsonString: String?) {
+    internal func sendJsonToMiniApp(string jsonString: String?) {
         NotificationCenter.default.sendCustomEvent(
             MiniAppEvent.Event(
                 miniAppId: appId,