+++ Tests/Unit/RealMiniAppViewTests.swift
@@ -427,17 +427,17 @@ class RealMiniAppUniversalBridgeTests: QuickSpec {
                     miniAppView.messageBodies = []
                 }
                 it("will post miniappRecieveJsonString event with empty string") {
-                    miniAppView.sendJsonStringToMiniApp(string: "")
+                    miniAppView.sendJsonToMiniApp(string: "")
                     expect(miniAppView.messageBodies.count).toEventually(equal(1))
                     expect(miniAppView.messageBodies[0]).toEventually(contain(MiniAppEvent.miniappRecieveJsonString.rawValue))
                 }
                 it("will send json string content to mini app") {
-                    miniAppView.sendJsonStringToMiniApp(string: "{\"action\":\"sendJsonToHostapp\",\"param\":{\"jsonInfo\":{\"content\":\"{\\\"data\\\":\\\"Thisisasamplejsoninformation\\\"}\"}},\"id\":\"10.822978364672421\"}")
+                    miniAppView.sendJsonToMiniApp(string: "{\"action\":\"sendJsonToHostapp\",\"param\":{\"jsonInfo\":{\"content\":\"{\\\"data\\\":\\\"Thisisasamplejsoninformation\\\"}\"}},\"id\":\"10.822978364672421\"}")
                     expect(miniAppView.messageBodies.count).toEventually(equal(1))
                     expect(miniAppView.messageBodies[0]).toEventually(contain(MiniAppEvent.miniappRecieveJsonString.rawValue))
                 }
                 it("will send string content to mini app") {
-                    miniAppView.sendJsonStringToMiniApp(string: "Some string content")
+                    miniAppView.sendJsonToMiniApp(string: "Some string content")
                     expect(miniAppView.messageBodies.count).toEventually(equal(1))
                     expect(miniAppView.messageBodies[0]).toEventually(contain(MiniAppEvent.miniappRecieveJsonString.rawValue))
                 }